patch: Fix broken test. Add some new tests. Move tests.
